TextTranslation/EnglishToMultiLanguageTranslation/requirements.txt
google-cloud-translate==3.1.0